@@ -33,7 +33,7 @@ setup(
     url="https://github.com/kscalelabs/urdf2mjcf",
     long_description=long_description,
     long_description_content_type="text/markdown",
-    python_requires=">=3.8",
+    python_requires=">=3.11",
     install_requires=requirements,
     tests_require=requirements_dev,
     extras_require={"dev": requirements_dev},
